Massive amount of work to get it working in Docker Compose
I've blacklisted the QikProp directory for now, so it should not be
added to the Git, and we will want to move it to somewhere more secure
than the git eventually.

Still need to clean up the interface to allow just basic info without
overloading users, but overall works at its core level.