Fix hashing for input files.
Original implementation relied on the `request` object which did not function properly with the WTForm fields. New implementation does a full `read()` of the file and parses it through SHA-1 before resetting the `seek(0)` and returning the checksum.

Files are now properly hashed and staged and not just the hash of "".

Update version number to 0.2